fix: code review fixes
